python/acontext-basic/pyproject.toml
[project]
name = "acontext-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio",
    "acontext",
    "python-dotenv",
]
